+++ apps/io-services-cms-webapp/src/main.ts
@@ -1,15 +1,30 @@
+import {
+  ServiceLifecycle,
+  ServicePublication,
+  stores,
+} from "@io-services-cms/models";
+import * as O from "fp-ts/Option";
+import * as RA from "fp-ts/ReadonlyArray";
+import * as RR from "fp-ts/ReadonlyRecord";
 import { pipe } from "fp-ts/lib/function";
-import { ServiceLifecycle, stores } from "@io-services-cms/models";
-import { createWebServer } from "./webservice";
-import { expressToAzureFunction } from "./lib/azure/adapters";
 import { getConfigOrThrow } from "./config";
-import { getApimClient } from "./lib/clients/apim-client";
+import {
+  expressToAzureFunction,
+  toAzureFunctionHandler,
+} from "./lib/azure/adapters";
 import { getDatabase } from "./lib/azure/cosmos";
+import { getApimClient } from "./lib/clients/apim-client";
 import { jiraClient } from "./lib/clients/jira-client";
 import { createRequestReviewHandler } from "./reviewer/request-review-handler";
+import { createReviewCheckerHandler } from "./reviewer/review-checker-handler";
 import { apimProxy } from "./utils/apim-proxy";
-import { getDao } from "./utils/service-review-dao";
 import { jiraProxy } from "./utils/jira-proxy";
+import { createWebServer } from "./webservice";
+
+import { processBatchOf, setBindings } from "./lib/azure/misc";
+import { handler as onServiceLifecycleChangeHandler } from "./watchers/on-services-lifecycles-change";
+
+import { getDao } from "./utils/service-review-dao";
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires, @typescript-eslint/no-unused-vars
 const BASE_PATH = require("../host.json").extensions.http.routePrefix;
@@ -29,6 +44,12 @@ const serviceLifecycleStore = stores.createCosmosStore(
   ServiceLifecycle.ItemType
 );
 
+// create a store for the ServicePublication finite state machine
+const servicePublicationStore = stores.createCosmosStore(
+  cosmos.container(config.COSMOSDB_CONTAINER_SERVICE_PUBBLICATIONS),
+  ServicePublication.ItemType
+);
+
 // entrypoint for all http functions
 export const httpEntryPoint = pipe(
   {
@@ -36,12 +57,13 @@ export const httpEntryPoint = pipe(
     apimClient,
     config,
     serviceLifecycleStore,
+    servicePublicationStore,
   },
   createWebServer,
   expressToAzureFunction
 );
 
-export const queueEntryPoint = createRequestReviewHandler(
+export const createRequestReviewEntryPoint = createRequestReviewHandler(
   getDao(config),
   jiraProxy(jiraClient(config)),
   apimProxy(
@@ -49,4 +71,24 @@ export const queueEntryPoint = createRequestReviewHandler(
     config.AZURE_APIM_RESOURCE_GROUP,
     config.AZURE_APIM
   )
+);
+
+export const serviceReviewCheckerEntryPoint = createReviewCheckerHandler(
+  getDao(config),
+  jiraProxy(jiraClient(config)),
+  serviceLifecycleStore
+);
+
+export const onServiceLifecycleChangeEntryPoint = pipe(
+  onServiceLifecycleChangeHandler,
+  processBatchOf(ServiceLifecycle.ItemType),
+  setBindings((results) => ({
+    requestReview: pipe(
+      results,
+      RA.map(RR.lookup("requestReview")),
+      RA.filter(O.isSome),
+      RA.map((item) => pipe(item.value, JSON.stringify))
+    ),
+  })),
+  toAzureFunctionHandler
 );