+++ apps/io-services-cms-webapp/src/webservice/__tests__/index.test.ts
@@ -1,17 +1,22 @@
-import { describe, it, expect, vi, afterEach } from "vitest";
-import request from "supertest";
-import * as TE from "fp-ts/lib/TaskEither";
-import * as O from "fp-ts/lib/Option";
-import { ServiceLifecycle, stores } from "@io-services-cms/models";
-import { createWebServer } from "../index";
 import { ApiManagementClient } from "@azure/arm-apimanagement";
-import { IConfig } from "../../config";
-import { UserGroup } from "@pagopa/io-functions-commons/dist/src/utils/middlewares/azure_api_auth";
 import {
-  getUserByEmail,
+  ServiceLifecycle,
+  ServicePublication,
+  stores,
+} from "@io-services-cms/models";
+import { UserGroup } from "@pagopa/io-functions-commons/dist/src/utils/middlewares/azure_api_auth";
+import * as O from "fp-ts/lib/Option";
+import * as TE from "fp-ts/lib/TaskEither";
+import request from "supertest";
+import { afterEach, describe, expect, it, vi } from "vitest";
+import { IConfig } from "../../config";
+import {
   getProductByName,
+  getUserByEmail,
   upsertSubscription,
 } from "../../lib/clients/apim-client";
+import { createWebServer } from "../index";
+import { itemToResponse as getPublicationItemToResponse } from "../controllers/get-service-publication/converters";
 
 vi.mock("../../lib/clients/apim-client", async () => {
   const anApimResource = { id: "any-id", name: "any-name" };
@@ -27,6 +32,9 @@ vi.mock("../../lib/clients/apim-client", async () => {
 const serviceLifecycleStore =
   stores.createMemoryStore<ServiceLifecycle.ItemType>();
 
+const servicePublicationStore =
+  stores.createMemoryStore<ServicePublication.ItemType>();
+
 const mockApimClient = {} as unknown as ApiManagementClient;
 const mockConfig = {} as unknown as IConfig;
 
@@ -40,6 +48,7 @@ describe("WebService", () => {
     apimClient: mockApimClient,
     config: mockConfig,
     serviceLifecycleStore,
+    servicePublicationStore,
   });
 
   describe("createService", () => {
@@ -148,6 +157,276 @@ describe("WebService", () => {
 
       expect(response.statusCode).toBe(500);
       expect(spied).not.toHaveBeenCalled();
+    });
+  });
+
+  describe("reviewService", () => {
+    const aService = {
+      id: "aServiceId",
+      data: {
+        name: "aServiceName",
+        description: "aServiceDescription",
+        authorized_recipients: [],
+        max_allowed_payment_amount: 123,
+        metadata: {
+          address: "via tal dei tali 123",
+          email: "service@email.it",
+          pec: "service@pec.it",
+          scope: "LOCAL",
+        },
+        organization: {
+          name: "anOrganizationName",
+          fiscal_code: "12345678901",
+        },
+        require_secure_channel: false,
+      },
+    } as unknown as ServiceLifecycle.ItemType;
+
+    it("should fail when cannot find requested service", async () => {
+      const response = await request(app)
+        .put("/api/services/s1/review")
+        .send()
+        .set("x-user-email", "example@email.com")
+        .set("x-user-groups", UserGroup.ApiServiceWrite)
+        .set("x-user-id", "any-user-id")
+        .set("x-subscription-id", "any-subscription-id");
+
+      expect(response.statusCode).toBe(500); // FIXME: should be 404 (or 409)
+    });
+
+    it("should fail when requested operation in not allowed (transition's preconditions fails)", async () => {
+      serviceLifecycleStore.save("s1", {
+        ...aService,
+        fsm: { state: "approved" },
+      });
+
+      const response = await request(app)
+        .put("/api/services/s1/review")
+        .send()
+        .set("x-user-email", "example@email.com")
+        .set("x-user-groups", UserGroup.ApiServiceWrite)
+        .set("x-user-id", "any-user-id")
+        .set("x-subscription-id", "any-subscription-id");
+
+      expect(response.statusCode).toBe(500); // FIXME: should be 409
+    });
+
+    it("should not allow the operation without right group", async () => {
+      const response = await request(app)
+        .put("/api/services/s1/review")
+        .send()
+        .set("x-user-email", "example@email.com")
+        .set("x-user-groups", "OtherGroup")
+        .set("x-user-id", "any-user-id")
+        .set("x-subscription-id", "any-subscription-id");
+
+      expect(response.statusCode).toBe(403);
+    });
+
+    it("should submit a service", async () => {
+      serviceLifecycleStore.save("s1", {
+        ...aService,
+        fsm: { state: "draft" },
+      });
+
+      const response = await request(app)
+        .put("/api/services/s1/review")
+        .send()
+        .set("x-user-email", "example@email.com")
+        .set("x-user-groups", UserGroup.ApiServiceWrite)
+        .set("x-user-id", "any-user-id")
+        .set("x-subscription-id", "any-subscription-id");
+
+      expect(response.statusCode).toBe(204);
+    });
+  });
+
+  const aService = {
+    id: "aServiceId",
+    data: {
+      name: "aServiceName",
+      description: "aServiceDescription",
+      authorized_recipients: [],
+      max_allowed_payment_amount: 123,
+      metadata: {
+        address: "via tal dei tali 123",
+        email: "service@email.it",
+        pec: "service@pec.it",
+        scope: "LOCAL",
+      },
+      organization: {
+        name: "anOrganizationName",
+        fiscal_code: "12345678901",
+      },
+      require_secure_channel: false,
+    },
+  } as unknown as ServicePublication.ItemType;
+
+  describe("publishService", () => {
+    it("should fail when cannot find requested service", async () => {
+      const response = await request(app)
+        .post("/api/services/s1/release")
+        .send()
+        .set("x-user-email", "example@email.com")
+        .set("x-user-groups", UserGroup.ApiServiceWrite)
+        .set("x-user-id", "any-user-id")
+        .set("x-subscription-id", "any-subscription-id");
+
+      expect(response.statusCode).toBe(500); // FIXME: should be 404 (or 409)
+    });
+
+    it("should fail when requested operation in not allowed (transition's preconditions fails)", async () => {
+      servicePublicationStore.save("s1", {
+        ...aService,
+        fsm: { state: "published" },
+      });
+
+      const response = await request(app)
+        .post("/api/services/s1/release")
+        .send()
+        .set("x-user-email", "example@email.com")
+        .set("x-user-groups", UserGroup.ApiServiceWrite)
+        .set("x-user-id", "any-user-id")
+        .set("x-subscription-id", "any-subscription-id");
+
+      expect(response.statusCode).toBe(500); // FIXME: should be 409
+    });
+
+    it("should not allow the operation without right group", async () => {
+      const response = await request(app)
+        .post("/api/services/s1/release")
+        .send()
+        .set("x-user-email", "example@email.com")
+        .set("x-user-groups", "OtherGroup")
+        .set("x-user-id", "any-user-id")
+        .set("x-subscription-id", "any-subscription-id");
+
+      expect(response.statusCode).toBe(403);
+    });
+
+    it("should publish a service", async () => {
+      servicePublicationStore.save("s1", {
+        ...aService,
+        fsm: { state: "unpublished" },
+      });
+
+      const response = await request(app)
+        .post("/api/services/s1/release")
+        .send()
+        .set("x-user-email", "example@email.com")
+        .set("x-user-groups", UserGroup.ApiServiceWrite)
+        .set("x-user-id", "any-user-id")
+        .set("x-subscription-id", "any-subscription-id");
+
+      expect(response.statusCode).toBe(204);
+    });
+  });
+
+  describe("getServicePublication", () => {
+    it("should fail when cannot find requested service", async () => {
+      const response = await request(app)
+        .delete("/api/services/s3/release")
+        .send()
+        .set("x-user-email", "example@email.com")
+        .set("x-user-groups", UserGroup.ApiServiceWrite)
+        .set("x-user-id", "any-user-id")
+        .set("x-subscription-id", "any-subscription-id");
+
+      expect(response.statusCode).toBe(500); // FIXME: should be 404 (or 409)
+    });
+
+    const asServiceWithStatus = {
+      ...aService,
+      fsm: { state: "published" },
+    } as unknown as ServicePublication.ItemType;
+
+    it("should retrieve a service", async () => {
+      servicePublicationStore.save("s3", asServiceWithStatus);
+
+      const response = await request(app)
+        .get("/api/services/s3/release")
+        .send()
+        .set("x-user-email", "example@email.com")
+        .set("x-user-groups", UserGroup.ApiServiceWrite)
+        .set("x-user-id", "any-user-id")
+        .set("x-subscription-id", "any-subscription-id");
+
+      expect(JSON.stringify(response.body)).toBe(
+        JSON.stringify(getPublicationItemToResponse(asServiceWithStatus))
+      );
+      expect(response.statusCode).toBe(200);
+    });
+
+    it("should not allow the operation without right group", async () => {
+      const response = await request(app)
+        .get("/api/services/s3/release")
+        .send()
+        .set("x-user-email", "example@email.com")
+        .set("x-user-groups", "OtherGroup")
+        .set("x-user-id", "any-user-id")
+        .set("x-subscription-id", "any-subscription-id");
+
+      expect(response.statusCode).toBe(403);
+    });
+  });
+
+  describe("unPublishService", () => {
+    it("should fail when cannot find requested service", async () => {
+      const response = await request(app)
+        .delete("/api/services/s2/release")
+        .send()
+        .set("x-user-email", "example@email.com")
+        .set("x-user-groups", UserGroup.ApiServiceWrite)
+        .set("x-user-id", "any-user-id")
+        .set("x-subscription-id", "any-subscription-id");
+
+      expect(response.statusCode).toBe(500); // FIXME: should be 404 (or 409)
+    });
+
+    it("should fail when requested operation in not allowed (transition's preconditions fails)", async () => {
+      servicePublicationStore.save("s2", {
+        ...aService,
+        fsm: { state: "unpublished" },
+      });
+
+      const response = await request(app)
+        .delete("/api/services/s2/release")
+        .send()
+        .set("x-user-email", "example@email.com")
+        .set("x-user-groups", UserGroup.ApiServiceWrite)
+        .set("x-user-id", "any-user-id")
+        .set("x-subscription-id", "any-subscription-id");
+
+      expect(response.statusCode).toBe(500); // FIXME: should be 409
+    });
+
+    it("should not allow the operation without right group", async () => {
+      const response = await request(app)
+        .delete("/api/services/s2/release")
+        .send()
+        .set("x-user-email", "example@email.com")
+        .set("x-user-groups", "OtherGroup")
+        .set("x-user-id", "any-user-id")
+        .set("x-subscription-id", "any-subscription-id");
+
+      expect(response.statusCode).toBe(403);
+    });
+
+    it("should unpublish a service", async () => {
+      servicePublicationStore.save("s2", {
+        ...aService,
+        fsm: { state: "published" },
+      });
+
+      const response = await request(app)
+        .delete("/api/services/s2/release")
+        .send()
+        .set("x-user-email", "example@email.com")
+        .set("x-user-groups", UserGroup.ApiServiceWrite)
+        .set("x-user-id", "any-user-id")
+        .set("x-subscription-id", "any-subscription-id");
+
+      expect(response.statusCode).toBe(204);
     });
   });
 });